+++ changelog.md
@@ -3,10 +3,10 @@ v2.0.0 (2018-02-21)
 
 * Improved Data Source 
 * Feature - Search Methods, Classes, and Modules
-* Feature - Focus Highlight on Classes and Modules in Search Matches
+* Feature - Focus Highlight on Classes & Modules in Search Matches
 * Feature - Display Syntax
 * Feature - Display Source Code
-* Added Learn More Option in Main Menu
+* Added Learn More Option in Main Menu with Changelog & Coming Soon
 
 v1.1.5 (2018-02-01)
 ------------------